wlan: call wpa_supplicant DISCONNECT
wpa_supplicant (or any tool really) can only scan hidden APs by actively
including the SSID name in the scan. This can be done by 'adding' the
network to wpa_supplicant's list of networks (much like having
network={} in /etc/wpa_supplicant/wpa_supplicant.conf). The problem is
that is only actually scans for them when the networks are not disabled,
but if they are enabled, then wpa supplicant typically just connects to
them automatically, which it should really not do in our case.

So then there is a special command in wpa_cli (or really, the
wpa_supplicant control socket interface) - DISCONNECT, this prevents
wpa_supplicant from connecting again until it's actively told to
connect. This allows having hidden AP networks "enabled" without
wpa_supplicant connecting to them. And of course, if we are already
connected to something else, then we don't need to do this, it will just
work (tm).

The dbus interface also has fi.w1.wpa_supplicant1.Interface.Disconnect,
but this unfortunately does not have the same effect. So we should
probably raise that on the wpa_supplicant mailing list, and meanwhile
also implement the socket control interface in
libicd-network-wpasupplicant (next to the dbus interface...).