Switch to new zero based interface path
The new wpa_supplicant uses a zero based interface index.

This code was hacky to begin with, we probably need to interface over
the existing interfaces, or get a path like this:

mdbus2 -s  fi.w1.wpa_supplicant1 /fi/w1/wpa_supplicant1 fi.w1.wpa_supplicant1.GetInterface "wlan0"

This will require all the wpaicd.c code to accept some data
containing at least the interface name, and it also means we have to
listen for InterfaceAdded and InterfaceRemoved signals, further
complicating the code tha was meant to be simple.

We should fix this in the future, but that is not yet now.