start assoc timer on 'disconnected' event
we may receive disconnected event on every stage, however, if we get it
before we are associating, we will never time-out or close connection.

Fix that by (re)starting assoc timer on 'disconnected' event, so if
connection is not established within a reasonable timeout we report error
to the upper layers.

Clean-up some duplicated code while at it.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>